add filtering of unused supertypes
