+++ plugins/groovy/src/org/jetbrains/plugins/groovy/intentions/style/inference/graph/InferenceUnitGraphBuilder.kt
@@ -1,14 +1,14 @@
 // Copyright 2000-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package org.jetbrains.plugins.groovy.intentions.style.inference.graph
 
-import com.intellij.psi.PsiClassType
-import com.intellij.psi.PsiIntersectionType
-import com.intellij.psi.PsiType
-import com.intellij.psi.PsiTypeVisitor
+import com.intellij.psi.*
 import com.intellij.psi.impl.source.resolve.graphInference.InferenceBound
 import com.intellij.psi.impl.source.resolve.graphInference.InferenceVariable
 import com.intellij.util.containers.BidirectionalMap
+import org.jetbrains.plugins.groovy.intentions.style.inference.ensureWildcards
 import org.jetbrains.plugins.groovy.intentions.style.inference.flattenIntersections
+import org.jetbrains.plugins.groovy.lang.psi.GroovyPsiElementFactory
+import org.jetbrains.plugins.groovy.lang.psi.util.GroovyCommonClassNames
 import org.jetbrains.plugins.groovy.lang.resolve.processors.inference.GroovyInferenceSession
 import org.jetbrains.plugins.groovy.lang.resolve.processors.inference.type
 
@@ -25,15 +25,30 @@ class InferenceUnitGraphBuilder {
                                           session: GroovyInferenceSession,
                                           flexibleTypes: Collection<PsiType> = emptySet(),
                                           constantTypes: Collection<PsiType> = emptySet(),
-                                          forbiddingTypes: Collection<PsiType> = emptySet()): InferenceUnitGraph {
+                                          forbiddingTypes: Collection<PsiType> = emptySet(),
+                                          appearedClassTypes: Map<String, List<PsiClass?>>): InferenceUnitGraph {
       val map = BidirectionalMap<InferenceUnit, InferenceVariable>()
       val builder = InferenceUnitGraphBuilder()
       for (variable in variables) {
         val typeParameter = variable.parameter.type()
+        val parameterTypes = variable.parameter.extendsList.referencedTypes.filter {
+          appearedClassTypes[typeParameter.className]?.contains(it.resolve()) ?: false
+        }
+        val instantiation = variable.instantiation
+        val partialInstantiation = when {
+          instantiation is PsiIntersectionType &&
+          instantiation.conjuncts.map { it.canonicalText }.contains(GroovyCommonClassNames.GROOVY_OBJECT)
+          && instantiation.conjuncts.size == 2 -> (instantiation.conjuncts.asIterable() - (instantiation.conjuncts.find { it.canonicalText == GroovyCommonClassNames.GROOVY_OBJECT })).first()
+          else -> instantiation
+        }
         val validType = when {
-          variable.parameter.extendsList.referencedTypes.size > 1 -> PsiIntersectionType.createIntersection(
-            variable.parameter.extendsList.referencedTypes.toList())
-          else -> variable.parameter.extendsList.referencedTypes.firstOrNull() ?: variable.instantiation
+          parameterTypes.size > 1 -> PsiIntersectionType.createIntersection(
+            parameterTypes.toList())
+          // questionable condition. I guess, we can allow LUB instead of Object if variable will instantiate into a parametrized type.
+          parameterTypes.isEmpty() && partialInstantiation is PsiIntersectionType -> PsiType.getJavaLangObject(session.manager,
+                                                                                                                 session.scope)
+          else -> parameterTypes.firstOrNull() ?: partialInstantiation!!.ensureWildcards(
+            GroovyPsiElementFactory.getInstance(variable.project), variable.manager)
         }
         val core = InferenceUnit(variable.parameter,
                                  typeParameter in flexibleTypes,
@@ -45,6 +60,7 @@ class InferenceUnitGraphBuilder {
         }
         map[core] = variable
       }
+
       val entries = map.entries.sortedBy { (unit, _) -> unit.toString() }
       for ((unit, variable) in entries) {
         deepConnect(session, map, (variable.getBounds(InferenceBound.UPPER) + variable.getBounds(InferenceBound.EQ))
@@ -75,7 +91,7 @@ class InferenceUnitGraphBuilder {
       }
       for (type in typeList.flattenIntersections()) {
         val dependentVariable = session.getInferenceVariable(type)
-        if (dependentVariable != null) {
+        if (dependentVariable != null && dependentVariable in map.values) {
           relationHandler(map.getKeysByValue(dependentVariable)!!.first())
         }
         else {
@@ -135,7 +151,7 @@ class InferenceUnitGraphBuilder {
       superTypesMap.add(mutableSetOf())
       subTypesMap.add(mutableSetOf())
     }
-    val unitIndexMap : Map<InferenceUnit, Int> = registeredUnits.zip(registeredUnits.indices).toMap()
+    val unitIndexMap: Map<InferenceUnit, Int> = registeredUnits.zip(registeredUnits.indices).toMap()
     for ((left, right, type) in relations) {
       when (type) {
         RelationType.SUPER -> subTypesMap[unitIndexMap.getValue(left)].add { inferenceNodes[unitIndexMap.getValue(right)] }
