+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/DeepEqualityOfWildcards_after.groovy
@@ -1,4 +1,4 @@
-def <U0 extends java.util.List<T0>, T0> Object foo(List<U0> a, List<? extends U0> b) {
+def <U0 extends java.util.List<?>> Object foo(List<U0> a, List<? extends U0> b) {
   a.add(b.get(0))
 }
