+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ParametrizedInterface2.groovy
@@ -1,0 +1,8 @@
+def foo(a, b) {
+  a.compareTo(b)
+}
+
+def m() {
+  foo(1, 2)
+  foo('s', 'q')
+}
