+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ParametrizedInterface2_after.groovy
@@ -1,0 +1,8 @@
+def <U0> Object foo(Comparable<U0> a, U0 b) {
+  a.compareTo(b)
+}
+
+def m() {
+  foo(1, 2)
+  foo('s', 'q')
+}
