+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ParametrizedInterface_after.groovy
@@ -1,4 +1,4 @@
-def <W0 extends java.io.Serializable & java.lang.Comparable<? extends U0>, U0 extends java.io.Serializable & java.lang.Comparable<?>> void foo(W0 a) {
+void foo(Object a) {
 }
 
 def m() {
