+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ParametrizedMethod.groovy
@@ -1,0 +1,15 @@
+def foo(a, b) {
+  a.tie(b)
+}
+
+
+class A{
+  def <T> void tie(T t) {}
+}
+
+class B {}
+
+
+def m(A a, B b) {
+  foo(a, b)
+}
