+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/ThreeSupertypes_after.groovy
@@ -1,4 +1,4 @@
-def <V0 extends groovy.lang.GroovyObject> Object foo(C<V0, V0, V0> a, V0 b) {
+def <V0> Object foo(C<V0, V0, V0> a, V0 b) {
   a.doU(b)
   a.doT(b)
   a.doV(b)
