+++ plugins/groovy/testdata/refactoring/inferMethodParametersTypes/TwoSupertypes_after.groovy
@@ -1,4 +1,4 @@
-def <U0 extends groovy.lang.GroovyObject> Object foo(C<U0, U0> a, U0 b) {
+def <U0> Object foo(C<U0, U0> a, U0 b) {
   a.doT(b)
   a.doU(b)
 }
